Create placeQuant.html
d3 bubble pack-style data visualization with a twist - pulls directly from database via the associated php script - displays place names from a library catalog - size of bubble is based on number of records with that place name. Clicking on the bubble opens a search for that place name in the catalog in a new browser window.